refactor(trainer): share one async batch loop between AR and diffusion
_sync_rollout, _next_rollout_batch, _submit_generations,
_build_async_sample, _score_completed, and the train() loop were
byte-identical (or 8 lines apart) across AsyncARTrainer and
AsyncDiffusionTrainer. Fold them into AsyncRolloutTrainerMixin in
trainer/async_rollout.py; the trainers keep only their domain hooks
(_advantage_and_train, _async_wandb_extra, _boundary_evaluate) and
their public train() signatures. RolloutManager still solely owns the
published rollout version.
