+++ unirl/trainer/async_diffusion.py
@@ -7,31 +7,19 @@ state. A completed batch is scored before its replacement is submitted.
 
 from __future__ import annotations
 
-import logging
 import time
-from typing import Any, Optional, Tuple
+from typing import Any, Dict, Optional, Tuple
 
 import torch
 
 from unirl.distributed.tensor import hydrate
-from unirl.rollout.manager import RolloutManager, keep_within_lag
 from unirl.train.stack import TrainStepResult
-from unirl.trainer.async_rollout import (
-    boundary_launch_slots,
-    combine_rollout_chunks,
-    next_hard_boundary,
-    rollout_version_metrics,
-    training_version_metrics,
-)
-from unirl.trainer.base import unwrap_replicated_int
+from unirl.trainer.async_rollout import AsyncRolloutTrainerMixin, training_version_metrics
 from unirl.trainer.diffusion import DiffusionTrainer
 from unirl.types.sample import Sample
-from unirl.types.sampling import total_samples_per_prompt
-
-logger = logging.getLogger(__name__)
 
 
-class AsyncDiffusionTrainer(DiffusionTrainer):
+class AsyncDiffusionTrainer(AsyncRolloutTrainerMixin, DiffusionTrainer):
     """Disaggregated async diffusion trainer (two slabs, resident engine, cross-slab sync)."""
 
     def __init__(
@@ -67,15 +55,6 @@ class AsyncDiffusionTrainer(DiffusionTrainer):
             raise ValueError(f"num_updates_per_batch must be >= 1, got {self._num_updates_per_batch}")
         self._train_version = 0
         self._batches_since_sync = 0
-
-    def _build_async_sample(self, gen_id: int) -> Sample:
-        """Consume one data batch and build the request Sample for ``gen_id``."""
-        return self._build_request_sample(self.data_source.get_samples(self.batch_size), gen_id)
-
-    def _score_completed(self, gen_id: int, completed: Sample) -> Sample:
-        scored = self.reward.score_and_attach(completed)
-        self._drop_decoded(scored, rollout_id=gen_id)
-        return scored
 
     def _advantage_and_train(
         self,
@@ -129,152 +108,16 @@ class AsyncDiffusionTrainer(DiffusionTrainer):
         load_dir: Optional[str] = None,
         save_mode: str = "auto",
     ) -> None:
-        start_rollout = self.maybe_load_checkpoint(load_dir, num_rollouts=num_rollouts)
-        resumed = bool(load_dir)
-        self._train_version = unwrap_replicated_int(
-            self.backend.get_optimizer_step_count(),
-            name="backend optimizer step count",
-        )
-        self._batches_since_sync = 0
-        for _ in range(start_rollout):
-            self.data_source.get_samples(self.batch_size)
-        staleness_budget = (self._weight_sync_interval - 1) * self._num_updates_per_batch
-        self._init_wandb(
+        self._train_async_loop(
             num_rollouts=num_rollouts,
-            extra={
-                "max_inflight": self._max_inflight,
-                "weight_sync_interval": self._weight_sync_interval,
-                "max_staleness": self._weight_sync_interval - 1,
-                "staleness_budget": staleness_budget,
-                "num_updates_per_batch": self._num_updates_per_batch,
-                "train_fraction": self._train_fraction,
-            },
+            save_interval=save_interval,
+            save_dir=save_dir,
+            load_dir=load_dir,
+            save_mode=save_mode,
         )
 
-        self._rollout_manager = RolloutManager(
-            self.rollout,
-            launchers=[lambda sample: self.rollout.launch_nowait("generate", sample)],
-            capacities=[self._max_inflight],
-            group_size=total_samples_per_prompt(self.sampling_params),
-            filter_fn=keep_within_lag(staleness_budget),
-        )
-        self._next_generation_id = start_rollout
+    def _async_wandb_extra(self) -> Dict[str, object]:
+        return {"train_fraction": self._train_fraction}
 
-        if resumed or self.eval_interval > 0:
-            self._sync_rollout(force=True, require_empty=True)
-        if self.eval_interval > 0:
-            self.evaluate(start_rollout, sync_weights=False, sleep_after=False)
-
-        try:
-            for rollout_id in range(start_rollout, num_rollouts):
-                t0 = time.perf_counter()
-                hard_boundary = next_hard_boundary(
-                    rollout_id,
-                    num_rollouts=num_rollouts,
-                    eval_interval=self.eval_interval,
-                    save_interval=save_interval,
-                )
-                sample, output_version = self._next_rollout_batch(
-                    rollout_id,
-                    num_rollouts=num_rollouts,
-                    hard_boundary=hard_boundary,
-                )
-                training_progress = rollout_id / max(1, num_rollouts - 1)
-                result, mean_reward = self._advantage_and_train(
-                    sample,
-                    training_progress=training_progress,
-                    rollout_id=rollout_id,
-                    t0=t0,
-                    extra_metrics=rollout_version_metrics(
-                        train_version=self._train_version,
-                        output_version=output_version,
-                        num_updates_per_batch=self._num_updates_per_batch,
-                    ),
-                )
-                self.wandb_logger.log_progress(rollout_id, num_rollouts, result, mean_reward, logger=logger)
-
-                step = rollout_id + 1
-                eval_due = self.eval_interval > 0 and step % self.eval_interval == 0
-                save_due = save_interval > 0 and (step % save_interval == 0 or step >= num_rollouts)
-                sync_due = step < num_rollouts and self._batches_since_sync >= self._weight_sync_interval
-                if eval_due or save_due or sync_due:
-                    self._sync_rollout(require_empty=eval_due or save_due)
-
-                if step >= num_rollouts and not self._rollout_manager.empty:
-                    raise RuntimeError("final rollout boundary requires an empty RolloutManager")
-
-                if eval_due:
-                    self.evaluate(step, sync_weights=False, sleep_after=False)
-                if save_due:
-                    self.maybe_save_checkpoint(
-                        rollout_id,
-                        num_rollouts,
-                        save_interval=save_interval,
-                        save_dir=save_dir,
-                        save_mode=save_mode,
-                    )
-        finally:
-            try:
-                self._rollout_manager.close()
-            finally:
-                self._finish_wandb()
-
-    def _sync_rollout(self, *, force: bool = False, require_empty: bool = False) -> None:
-        manager = self._rollout_manager
-        needs_publish = force or manager.published_version != self._train_version
-        if not needs_publish:
-            if require_empty and not manager.empty:
-                raise RuntimeError("eval/checkpoint boundary requires an empty RolloutManager")
-            self._batches_since_sync = 0
-            return
-
-        carried = manager.quiesce(current_version=self._train_version)
-        if require_empty and (carried or not manager.empty):
-            raise RuntimeError("eval/checkpoint boundary requires an empty RolloutManager")
-        if needs_publish:
-            manager.sync_weights(self.weight_sync, output_version=self._train_version)
-        if carried:
-            manager.submit(carried)
-        self._batches_since_sync = 0
-
-    def _next_rollout_batch(
-        self,
-        rollout_id: int,
-        *,
-        num_rollouts: int,
-        hard_boundary: int,
-    ) -> Tuple[Sample, int]:
-        manager = self._rollout_manager
-        inflight_count, ready_count = manager.counts
-        if inflight_count + ready_count == 0:
-            slots = boundary_launch_slots(
-                inflight_count=0,
-                ready_count=0,
-                max_inflight=self._max_inflight,
-                trained_batches=rollout_id,
-                num_rollouts=num_rollouts,
-                hard_boundary=hard_boundary,
-            )
-            self._submit_generations(slots)
-
-        groups = manager.collect(self.batch_size, current_version=self._train_version)
-        completed, gen_id, output_version = combine_rollout_chunks(groups)
-        scored = self._score_completed(gen_id, completed)
-
-        inflight_count, ready_count = manager.counts
-        slots = boundary_launch_slots(
-            inflight_count=inflight_count,
-            ready_count=ready_count + 1,
-            max_inflight=self._max_inflight,
-            trained_batches=rollout_id,
-            num_rollouts=num_rollouts,
-            hard_boundary=hard_boundary,
-        )
-        self._submit_generations(slots)
-        return scored, output_version
-
-    def _submit_generations(self, count: int) -> None:
-        for _ in range(count):
-            gen_id = self._next_generation_id
-            self._rollout_manager.submit([self._build_async_sample(gen_id)])
-            self._next_generation_id += 1
+    def _boundary_evaluate(self, rollout_id: int, *, initial: bool) -> None:
+        self.evaluate(rollout_id if initial else rollout_id + 1, sync_weights=False, sleep_after=False)
