+++ unirl/trainer/async_rollout.py
@@ -1,9 +1,19 @@
+"""Shared driver-side policy for the async batch trainers (AR and diffusion)."""
+
 from __future__ import annotations
 
-from typing import TYPE_CHECKING, List, Tuple
+import logging
+import time
+from typing import TYPE_CHECKING, Dict, List, Optional, Tuple
+
+from unirl.rollout.manager import RolloutManager, keep_within_lag
+from unirl.trainer.base import unwrap_replicated_int
+from unirl.types.sampling import total_samples_per_prompt
 
 if TYPE_CHECKING:
     from unirl.types.sample import Sample
+
+logger = logging.getLogger(__name__)
 
 
 def next_hard_boundary(
@@ -95,7 +105,197 @@ def _rollout_id(sample: "Sample") -> int:
     return int(next(iter(values)))
 
 
+class AsyncRolloutTrainerMixin:
+    """One async batch loop shared by ``AsyncARTrainer`` and ``AsyncDiffusionTrainer``.
+
+    The host trainer supplies the domain surface (``_build_request_sample``,
+    ``_drop_decoded``, ``_advantage_and_train``, checkpoint/wandb helpers, the
+    ``reward``/``backend``/``rollout``/``weight_sync`` remotes, and the
+    ``_max_inflight``/``_weight_sync_interval``/``_num_updates_per_batch``
+    knobs) plus the ``_async_wandb_extra`` / ``_boundary_evaluate`` hooks.
+    The mixin owns admission, boundary sync, and FIFO batch consumption;
+    ``RolloutManager`` alone owns the published rollout version.
+    """
+
+    def _async_wandb_extra(self) -> Dict[str, object]:
+        """Trainer-specific keys merged into the wandb run config."""
+        return {}
+
+    def _boundary_evaluate(self, rollout_id: int, *, initial: bool) -> None:
+        """Run the trainer's evaluation at a synced, empty rollout boundary."""
+        raise NotImplementedError
+
+    def _build_async_sample(self, gen_id: int) -> "Sample":
+        """Consume one data batch and build the request Sample for ``gen_id``."""
+        return self._build_request_sample(self.data_source.get_samples(self.batch_size), gen_id)
+
+    def _score_completed(self, gen_id: int, completed: "Sample") -> "Sample":
+        scored = self.reward.score_and_attach(completed)
+        self._drop_decoded(scored, rollout_id=gen_id)
+        return scored
+
+    def _train_async_loop(
+        self,
+        *,
+        num_rollouts: int,
+        save_interval: int,
+        save_dir: Optional[str],
+        load_dir: Optional[str],
+        save_mode: str,
+    ) -> None:
+        start_rollout = self.maybe_load_checkpoint(load_dir, num_rollouts=num_rollouts)
+        resumed = bool(load_dir)
+        self._train_version = unwrap_replicated_int(
+            self.backend.get_optimizer_step_count(),
+            name="backend optimizer step count",
+        )
+        self._batches_since_sync = 0
+        for _ in range(start_rollout):
+            self.data_source.get_samples(self.batch_size)
+        staleness_budget = (self._weight_sync_interval - 1) * self._num_updates_per_batch
+        self._init_wandb(
+            num_rollouts=num_rollouts,
+            extra={
+                "max_inflight": self._max_inflight,
+                "weight_sync_interval": self._weight_sync_interval,
+                "max_staleness": self._weight_sync_interval - 1,
+                "staleness_budget": staleness_budget,
+                "num_updates_per_batch": self._num_updates_per_batch,
+                **self._async_wandb_extra(),
+            },
+        )
+
+        self._rollout_manager = RolloutManager(
+            self.rollout,
+            launchers=[lambda sample: self.rollout.launch_nowait("generate", sample)],
+            capacities=[self._max_inflight],
+            group_size=total_samples_per_prompt(self.sampling_params),
+            filter_fn=keep_within_lag(staleness_budget),
+        )
+        self._next_generation_id = start_rollout
+
+        if resumed or self.eval_interval > 0:
+            self._sync_rollout(force=True, require_empty=True)
+        if self.eval_interval > 0:
+            self._boundary_evaluate(start_rollout, initial=True)
+
+        try:
+            for rollout_id in range(start_rollout, num_rollouts):
+                t0 = time.perf_counter()
+                hard_boundary = next_hard_boundary(
+                    rollout_id,
+                    num_rollouts=num_rollouts,
+                    eval_interval=self.eval_interval,
+                    save_interval=save_interval,
+                )
+                sample, output_version = self._next_rollout_batch(
+                    rollout_id,
+                    num_rollouts=num_rollouts,
+                    hard_boundary=hard_boundary,
+                )
+                training_progress = rollout_id / max(1, num_rollouts - 1)
+                result, mean_reward = self._advantage_and_train(
+                    sample,
+                    training_progress=training_progress,
+                    rollout_id=rollout_id,
+                    t0=t0,
+                    extra_metrics=rollout_version_metrics(
+                        train_version=self._train_version,
+                        output_version=output_version,
+                        num_updates_per_batch=self._num_updates_per_batch,
+                    ),
+                )
+                self.wandb_logger.log_progress(rollout_id, num_rollouts, result, mean_reward, logger=logger)
+
+                step = rollout_id + 1
+                eval_due = self.eval_interval > 0 and step % self.eval_interval == 0
+                save_due = save_interval > 0 and (step % save_interval == 0 or step >= num_rollouts)
+                sync_due = step < num_rollouts and self._batches_since_sync >= self._weight_sync_interval
+                if eval_due or save_due or sync_due:
+                    self._sync_rollout(require_empty=eval_due or save_due)
+
+                if step >= num_rollouts and not self._rollout_manager.empty:
+                    raise RuntimeError("final rollout boundary requires an empty RolloutManager")
+
+                if eval_due:
+                    self._boundary_evaluate(rollout_id, initial=False)
+                if save_due:
+                    self.maybe_save_checkpoint(
+                        rollout_id,
+                        num_rollouts,
+                        save_interval=save_interval,
+                        save_dir=save_dir,
+                        save_mode=save_mode,
+                    )
+        finally:
+            try:
+                self._rollout_manager.close()
+            finally:
+                self._finish_wandb()
+
+    def _sync_rollout(self, *, force: bool = False, require_empty: bool = False) -> None:
+        manager = self._rollout_manager
+        needs_publish = force or manager.published_version != self._train_version
+        if not needs_publish:
+            if require_empty and not manager.empty:
+                raise RuntimeError("eval/checkpoint boundary requires an empty RolloutManager")
+            self._batches_since_sync = 0
+            return
+
+        carried = manager.quiesce(current_version=self._train_version)
+        if require_empty and (carried or not manager.empty):
+            raise RuntimeError("eval/checkpoint boundary requires an empty RolloutManager")
+        if needs_publish:
+            manager.sync_weights(self.weight_sync, output_version=self._train_version)
+        if carried:
+            manager.submit(carried)
+        self._batches_since_sync = 0
+
+    def _next_rollout_batch(
+        self,
+        rollout_id: int,
+        *,
+        num_rollouts: int,
+        hard_boundary: int,
+    ) -> Tuple["Sample", int]:
+        manager = self._rollout_manager
+        inflight_count, ready_count = manager.counts
+        if inflight_count + ready_count == 0:
+            slots = boundary_launch_slots(
+                inflight_count=0,
+                ready_count=0,
+                max_inflight=self._max_inflight,
+                trained_batches=rollout_id,
+                num_rollouts=num_rollouts,
+                hard_boundary=hard_boundary,
+            )
+            self._submit_generations(slots)
+
+        groups = manager.collect(self.batch_size, current_version=self._train_version)
+        completed, gen_id, output_version = combine_rollout_chunks(groups)
+        scored = self._score_completed(gen_id, completed)
+
+        inflight_count, ready_count = manager.counts
+        slots = boundary_launch_slots(
+            inflight_count=inflight_count,
+            ready_count=ready_count + 1,
+            max_inflight=self._max_inflight,
+            trained_batches=rollout_id,
+            num_rollouts=num_rollouts,
+            hard_boundary=hard_boundary,
+        )
+        self._submit_generations(slots)
+        return scored, output_version
+
+    def _submit_generations(self, count: int) -> None:
+        for _ in range(count):
+            gen_id = self._next_generation_id
+            self._rollout_manager.submit([self._build_async_sample(gen_id)])
+            self._next_generation_id += 1
+
+
 __all__ = [
+    "AsyncRolloutTrainerMixin",
     "boundary_launch_slots",
     "combine_rollout_chunks",
     "next_hard_boundary",
